parkwise-fastapi/predictor/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
numpy==1.26.4
pandas==2.1.2
scikit-learn==1.3.2
xgboost==2.0.3
joblib==1.3.2
opencv-python==4.8.1.78
ultralytics==8.0.196
Pillow==10.0.1
torch==2.1.0
torchvision==0.16.0